Fixup cloning for ref HEAD
I noticed that passing HEAD as a ref is both sufficiently valid to clone
and fails our clone script with a cryptic exit code 1.

I think this is what is already meant to be guarded against here:

https://github.com/rwx-cloud/packages/blob/33af29f0246e71d29317e2fddf848cd05efda6da/git/clone/bin/git-clone#L180-L182

But the problem is that if grep doesn't match, it raises exit code 1
here:

https://github.com/rwx-cloud/packages/blob/33af29f0246e71d29317e2fddf848cd05efda6da/git/clone/bin/git-clone#L176

To even get the HEAD ref to show up in ls-remote, we'd need to do
something like:

```
git ls-remote origin
```

And even then, the actual value we'd want to match on is just `HEAD` and
not `refs/heads/HEAD`.

Anyway, this fixes it and adds a test.

Example failing run:

https://cloud.rwx.com/mint/rwx/runs/d9fb75e800f6437082af9bbf85eb24e7