Check out from get-latest-sha-for-ref's sha
There's a subtle race in our git/clone package today where:

- we resolve the latest sha for `${{ params.ref }}`
- if it changes, we bust the cache on `git-clone`
- but `git-clone` re-resolves the latest sha for  `${{ params.ref }}`

Meaning if another commit comes in between `get-latest-sha-for-ref`
busting the `git-clone` cache and `git-clone` actually doing its
checkout, the two tasks will resolve to different shas.

Practically, this doesn't matter. We're still appropriately cache
busting. Although in general, it is a nice property of CI/CD that it
operates on the same sha throughout the run.

However, I want to elevate some of the dynamic task generation logic
proposed in https://github.com/rwx-cloud/packages/pull/278.

If there's an earlier point when we know the resolved sha that we
can use to determine if a patch can be applied, `git-clone` can
continue to cache hit on the expensive cloning behavior.

So this closes that subtle race condition and in doing so, gives me an
earlier entrypoint to make patching decisions.

I've also done some renaming to make it more clear that this is no
longer just about cache busting.