Search entire terraform --version output
This assertion failed unexpectantly but when navigating the layer
filesystem, the terraform binary exists.

https://cloud.rwx.com/mint/rwx/tasks/0d977183461e8bf7badec5ae6f035676/layers/c9e0fb70-6bb9-4adb-9413-cc0bd6c5668a#path=%2Fusr%2Fbin

I'm changing the output and matching slightly because the output seems
to suggest that `head -n1` gave us an empty line:

https://cloud.rwx.com/mint/rwx/tasks/0d977183461e8bf7badec5ae6f035676/logs

So now we'll capture the whole multiline version output. And we'll match
against any line.