Enable auto-merge on update PRs
I approved several of these last week and forgot to click merge. If the
repo respects auto-merging, we should allow you to opt into enabling it
when the PR is created. This is acheived by first opening the PR and
then enabling auto-merge.

This will still block on any repo protection (build passing, approvals,
etc.) and will fail open if the repo in question doesn't have
auto-merging enabled.